+++ client/pages/calendar.tsx
@@ -1,16 +1,9 @@
-import { Layout } from "../components/layout";
-import { useAuthContext } from "../contexts/auth-context";
-import { useRouter } from "next/router";
-import { useEffect } from "react";
+import { PageWrapper } from "../components/page-wrapper";
 
 export default function Calendar() {
-  const auth = useAuthContext();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (!auth.user) {
-      router.replace("/");
-    }
-  });
-  return <Layout title="Calendar">LOGGED I</Layout>;
+  return (
+    <PageWrapper title="Calendar" isProtectedPage>
+      LOGGED I
+    </PageWrapper>
+  );
 }
