+++ client/pages/index.tsx
@@ -1,9 +1,9 @@
-import { Layout } from "../components/layout";
+import { PageWrapper } from "../components/page-wrapper";
 
 export default function Home() {
   return (
-    <Layout title="A simple calendar to keep track of your shifts">
+    <PageWrapper title="A simple calendar to keep track of your shifts">
       <img src="/images/illustration-calendar.svg" />
-    </Layout>
+    </PageWrapper>
   );
 }
