+++ client/pages/login.tsx
@@ -1,10 +1,11 @@
-import { Layout } from "../components/layout";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faEnvelope, faLock } from "@fortawesome/free-solid-svg-icons";
 import { Colors } from "../constants/colors";
 import { useForm } from "react-hook-form";
 import { useAuthContext } from "../contexts/auth-context";
 import { useRouter } from "next/router";
+import { api } from "../services/api";
+import { PageWrapper } from "../components/page-wrapper";
 
 type FormData = {
   email: string;
@@ -16,14 +17,15 @@ export default function Login() {
   const router = useRouter();
   const { register, handleSubmit, errors } = useForm<FormData>();
   const onSubmit = handleSubmit(async ({ email, password }) => {
-    auth.login({ email, password });
-    if (auth.token) {
-      await router.push("/calendar");
-    }
+    const response = await api.post(`login`, { email, password });
+
+    auth.setToken(response.data["jwt"] as string);
+
+    await router.push("/calendar");
   });
 
   return (
-    <Layout title="Login">
+    <PageWrapper title="Login">
       <h3 className="title is-3 has-text-primary">Login</h3>
 
       <div className="columns is-vcentered is-variable is-8">
@@ -83,6 +85,6 @@ export default function Login() {
           <img src="/images/illustration-login.svg" />
         </div>
       </div>
-    </Layout>
+    </PageWrapper>
   );
 }
